+++ rpg-p3-app/src/components/pages/Battlescreen/AttackBar/Attack.css
@@ -1,0 +1,4 @@
+#attackBox, #attackText{
+    justify-content: center;
+    align-items: center;
+}
