--- rpg-p3-app/src/components/pages/Battlescreen/AttackBar/TypingAttackBar.js
@@ -1,10 +1,0 @@
-// export default function AttackBar() {
-//   return (
-//     <section>
-//       <form>
-//         <label for="currentWord">Attack!!</label>
-//         <input type="text" id="currentWord" name="wordGame"/>
-//         <br/>
-//       </form>
-//     </section>
-//   )};
