+++ rpg-p3-app/src/components/pages/Battlescreen/AttackBar/index.js
@@ -1,0 +1,16 @@
+import './Attack.css'
+
+
+export default function AttackBar() {
+    return (
+        <section>
+            <Form.Label id="attackBox" htmlfor="attackBox">Placeholder text</Form.Label>
+            <Form.Control
+                type="text"
+            />
+            <Form.Text id="attackText" muted>
+                Get Ready to Attack!!!
+            </Form.Text>
+        </section>
+    )
+};
