+++ rpg-p3-app/src/components/pages/Battlescreen/Avatar/index.js
@@ -1,0 +1,10 @@
+import '../imgs/PlaceholderPlayer2';
+
+export default function Avatars(){
+    return (
+        <section className = "float-left">
+            <img src = {require('/imgs/PlaceholderPlayer2.png')}/>
+            <p>Player 1</p>
+        </section>
+    )
+};
