+++ rpg-p3-app/src/components/pages/Battlescreen/Enemy/Enemy.css
@@ -1,0 +1,4 @@
+.roundedCircle {
+    justify-content: center;
+    align-items: center;
+}
