--- rpg-p3-app/src/components/pages/Battlescreen/Enemy/Enemy.js
@@ -1,8 +1,0 @@
-// import Image from './imgs/PlaceholderPlayer2';
-// import 'Enemy.css';
-
-// export default function Enemy(){
-//     <section className = "roundedCircle">
-//             <img src = {require('../rpg-p3-app/imgs/PlaceholderPlayer2').default}/>
-//     </section>
-// }
