+++ rpg-p3-app/src/components/pages/Battlescreen/Enemy/index.js
@@ -1,0 +1,8 @@
+import '../imgs/PlaceholderPlayer1';
+import 'Enemy.css';
+
+export default function Enemy(){
+    <section className = "roundedCircle">
+            <img src = {require('../imgs/PlaceholderPlayer1')}/>
+    </section>
+}
