+++ rpg-p3-app/src/components/pages/Battlescreen/ProgressBar/Progess.css
@@ -1,0 +1,4 @@
+.hpBar {
+    justify-content: center;
+    align-items: center;
+}
