+++ rpg-p3-app/src/components/pages/Battlescreen/ProgressBar/index.js
@@ -1,11 +1,12 @@
-// import ProgressBar from 'react-bootstrap/ProgressBar';
+import ProgressBar from 'react-bootstrap/ProgressBar';
+import 'Progess.css';
 
-// export function AnimatedHP(){
-//     return (
-//         <section>
-//             <ProgressBar stripped variant = "danger" now = {100} ></ProgressBar>
-//             <ProgressBar animated now = {100} ></ProgressBar>
-//         </section>
-//     )}
+export function AnimatedHP(){
+    return (
+        <section id = "hpBar">
+            <ProgressBar stripped variant = "danger" now = {100} ></ProgressBar>
+            <ProgressBar animated now = {100} ></ProgressBar>
+        </section>
+    )}
 
-// export default AnimatedHP;
+export default AnimatedHP;
