+++ rpg-p3-app/src/components/pages/Battlescreen/imgs/PlaceholderPlayer2.png
